clean dependency in non-silo-tests
