+++ test/NonSiloTests/SerializationTests/ExternalSerializerTest.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using Orleans.Runtime;
 using Orleans.Serialization;
+using Tester.Serialization;
 using Xunit;
 
 namespace UnitTests.Serialization
@@ -59,54 +60,6 @@ namespace UnitTests.Serialization
         private class FakeSerializedWithNoCodegenSerializers : FakeSerialized
         {
             public string SomeMoreData;
-        }
-    }
-
-    public class FakeSerialized
-    {
-        public string SomeData;
-    }
-
-    public class FakeSerializer : IExternalSerializer
-    {
-        public static bool Initialized { get; set; }
-
-        public static bool IsSupportedTypeCalled { get; set; }
-
-        public static bool SerializeCalled { get; set; }
-
-        public static bool DeserializeCalled { get; set; }
-
-        public static bool DeepCopyCalled { get; set; }
-
-        public void Initialize(Logger logger)
-        {
-            Initialized = true;
-        }
-
-        public bool IsSupportedType(Type itemType)
-        {
-            IsSupportedTypeCalled = true;
-            return typeof(FakeSerialized).IsAssignableFrom(itemType);
-        }
-
-        public object DeepCopy(object source)
-        {
-            DeepCopyCalled = true;
-            return null;
-        }
-
-        public void Serialize(object item, BinaryTokenStreamWriter writer, Type expectedType)
-        {
-            SerializeCalled = true;
-            writer.WriteNull();
-        }
-
-        public object Deserialize(Type expectedType, BinaryTokenStreamReader reader)
-        {
-            DeserializeCalled = true;
-            reader.ReadToken();
-            return (FakeSerialized)Activator.CreateInstance(expectedType);
         }
     }
 }
