+++ vNext/src/Orleans/Properties/AssemblyInfo.cs
@@ -23,4 +23,5 @@ using System.Runtime.InteropServices;
 [assembly: InternalsVisibleTo("OrleansAzureUtils")]
 [assembly: InternalsVisibleTo("TesterInternal")]
 [assembly: InternalsVisibleTo("Tester.AzureUtils")]
+[assembly: InternalsVisibleTo("Tester")]
 [assembly: InternalsVisibleTo("Tester.SQLUtils")]
