+++ vNext/test/NonSiloTests/Orleans.NonSiloTests.csproj
@@ -1,7 +1,9 @@
-﻿<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" />
   <PropertyGroup Label="Configuration">
     <DefineConstants>NETSTANDARD;NETSTANDARD_TODO</DefineConstants>
+    <RootNamespace>Orleans.NonSiloTests</RootNamespace>
+    <AssemblyName>Orleans.NonSiloTests</AssemblyName>
   </PropertyGroup>
   <PropertyGroup>
     <TargetFramework>net462</TargetFramework>
